Template for fixing this problem. Need to use a helper
