+++ app/presenters/common_presenters.rb
@@ -12,24 +12,7 @@ module CommonPresenters
     list_of_fieldnames.map {|f| hl_field(document, f)}.reject {|x| x.empty?}.first || []
   end
 
-  def cit_xslt
-    load_xslt('CitOnly.xsl')
-  end
 
-  # @param [MiddleEnglishDictionary::Entry::Citation] cit The citation object
-  # @return [String, nil] The citatation transformed into HTML, or nil
-  def cit_html(cit)
-    rid = cit.bib.stencil.rid
-    if rid
-      bibid = Dromedary.hyp_to_bibid[rid.upcase] if rid
-      url   = bib_link_path(bibid)
-      xsl_transform_from_xml(cit.xml, cit_xslt, ["biburl", "'#{url}'"])
-    end
-
-  end
-
-  alias_method :cite_html, :cit_html
-  alias_method :citation_html, :cit_html
 
 
   # A convenience method to get the highlighted values for a field if
