+++ app/views/catalog/show_entry/_sense.html.erb
@@ -38,7 +38,7 @@
         <div class="citations">
           <ul>
             <% eg.citations.each do |cite| %>
-              <li class="citation-list-item"><%== presenter.citation_html(cite) %></li>
+              <li class="citation-list-item"><%== citation_html(cite) %></li>
             <% end %>
           </ul>
         </div>
