+++ config/routes.rb
@@ -4,9 +4,6 @@ Rails.application.routes.draw do
 
     mount Blacklight::Engine => Dromedary.config.relative_url_root
 
-    # Where should it look like it's mounted?
-    default_route = 'dictionary'
-
 
     # Splash pages
     match "dictionary/" => "catalog#home", as: :dictionary_home, via: [:get, :post], constraints: {query_string: ""}
@@ -16,29 +13,27 @@ Rails.application.routes.draw do
 
     # Rails doesn't allow dots in matched ids by default, because reasons.
     # Override the id matcher with an explicit constraint.
-    match "#{default_route}/(:id)/track" => 'catalog#show',
+    match "dictionary/(:id)/track" => 'catalog#show',
           :constraints                   => {:id => /[\p{Alnum}\-\.]+/}, via: [:get, :post]
 
-    match "bibliography/(:id)/track" => 'bibliography#show',
+    match "bibliography/(:id)/track" => 'bibliography#show', as: :bib_link,
           :constraints               => {:id => /(?:BIB|HYP)[T\d\-\.]+/i}, via: [:get, :post]
 
 
     match "bibliography/" => 'bibliography#index', via: [:get, :post]
 
-    match "bibliography/(:id)" => 'bibliography#show', as: :bib_link, via: [:get, :post],
-          constraints:         {:id => /\S\S+/}
 
 
     root to: "catalog#splash"
 
 
     # Force to go to root ('/'), not index.html
-    # get "/#{default_route}", to: redirect('/'), constraints: {query_string: ""}
+    # get "/dictionary", to: redirect('/'), constraints: {query_string: ""}
 
     concern :searchable, Blacklight::Routes::Searchable.new
 
 
-    resource :search, only: [:index], as: 'catalog', path: "/#{default_route}", controller: 'catalog' do
+    resource :search, only: [:index], as: 'catalog', path: "/dictionary", controller: 'catalog' do
       concerns :searchable
     end
 
@@ -54,7 +49,7 @@ Rails.application.routes.draw do
 
     concern :exportable, Blacklight::Routes::Exportable.new
 
-    resources :solr_documents, only: [:show], path: "/#{default_route}", controller: 'catalog' do
+    resources :solr_documents, only: [:show], path: "/dictionary", controller: 'catalog' do
       concerns :exportable
     end
